♻️ Split component files in extension generator (#32460)
* Split component files in extension generator

* First pass at updating docs

* Lint fixes

* Lint fixes
