+++ build-system/tasks/extension-generator/bento/amp-__component_name_hyphenated__/__component_version__/component.jss.js
