+++ contributing/building-a-bento-amp-extension.md
@@ -64,11 +64,12 @@ The directory structure is below:
 |   ├── test/                            # Element's unit test suite (req'd)
 |   |   ├── test-amp-my-element.js
 |   |   └── More test JS files (optional)
-|   ├── my-element.js                    # Preact implementation (req'd)
-|   ├── my-element.type.js               # Preact type definitions (req'd)
+|   ├── base-element.js                  # Preact base element (req'd)
+|   ├── component.js                     # Preact implementation (req'd)
+|   ├── component.type.js                # Preact type definitions (req'd)
 |   ├── amp-my-element.js                # Element's implementation (req'd)
 |   ├── amp-my-element.css               # Custom CSS (optional)
-|   ├── my-element.jss.js                # Preact JSS (optional)
+|   ├── component.jss.js                 # Preact JSS (optional)
 |   └── More JS files (optional)
 ├── validator-amp-my-element.protoascii  # Validator rules (req'd)
 ├── amp-my-element.md                    # Element's main documentation (req'd)
@@ -99,7 +100,7 @@ file (`extensions/amp-my-element/0.1/amp-my-element.js`). See [Experiments](#exp
 
 ```js
 import {func1, func2} from '../../../src/module';
-import {MyElement} from './my-element'; // Preact component.
+import {BaseElement} from './my-element'; // Preact base element.
 import {CSS} from '../../../build/amp-my-element-0.1.css';
 // more ES2015-style import statements.
 
@@ -109,7 +110,7 @@ const EXPERIMENT = 'amp-my-element';
 /** @const */
 const TAG = 'amp-my-element';
 
-class AmpMyElement extends AMP.PreactBaseElement {
+class AmpMyElement extends BaseElement {
   /** @override */
   init() {
     // Perform any additional processing of prop values that are not
@@ -130,19 +131,28 @@ class AmpMyElement extends AMP.PreactBaseElement {
   }
 }
 
-AmpMyElement['Component'] = MyElement;            // Component definition.
-
-AmpMyElement['props'] = {  // Map DOM attributes to Preact Component props.
-  'propName1': {attr: 'attr-name-1'},
-  'propName2': {attr: 'attr-name-2', type: 'number'},
-};
-
 AMP.extension('amp-my-element', '0.1', (AMP) => {
   AMP.registerElement('amp-my-element', AmpMyElement, CSS);
 });
 ```
 
-The following shows the corresponding overall structure of your Preact functional component implementation file (extensions/amp-my-element/0.1/my-element.js).
+The following shows the corresponding overall structure of your Preact base element implementation file (extensions/amp-my-element/0.1/base-element.js).
+
+```js
+import {MyElement} from './component'; // Preact component.
+import {PreactBaseElement} from '../../../src/preact/base-element'; // Preact component.
+
+export class BaseElement extends PreactBaseElement;
+
+BaseElement['Component'] = MyElement;            // Component definition.
+
+BaseElement['props'] = {  // Map DOM attributes to Preact Component props.
+  'propName1': {attr: 'attr-name-1'},
+  'propName2': {attr: 'attr-name-2', type: 'number'},
+};
+```
+
+The following shows the corresponding overall structure of your Preact functional component implementation file (extensions/amp-my-element/0.1/component.js).
 
 ```js
 import * as Preact from '../../../src/preact';
